requirements.txt
absl-py==0.9.0
aiobotocore==1.1.1
aiohttp==3.6.2
aioitertools==0.7.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
botocore==1.17.44
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
fsspec==0.8.3
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
Keras==2.4.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==4.7.6
networkx==2.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3fs==0.5.1
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.0
